pyproject.toml
[project]
name = "support-desk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.10.4",
    "python-dotenv>=1.2.2",
]
